requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pytesseract==0.3.10
pillow==10.0.0
python-multipart==0.0.6
